fix(codacy): remove util package findings
